requirements.txt
fastapi==0.109.2
mangum==0.17.0
PyNaCl==1.5.0
pydantic-settings==2.2.0
boto3